Split user into auth and profile
