+++ api/schema.py
@@ -1,12 +1,14 @@
 import graphene
 
-import api.schemas.user
+import api.schemas.auth
+import api.schemas.profile
 import api.schemas.presentation
 import api.schemas.conference
 import api.schemas.sponsor
 
 class Mutations(api.schemas.conference.Mutations,
-                api.schemas.user.Mutations,
+                api.schemas.auth.Mutations,
+                api.schemas.profile.Mutations,
                 api.schemas.presentation.Mutations,
                 api.schemas.sponsor.Mutations,
                 graphene.ObjectType):
@@ -14,7 +16,8 @@ class Mutations(api.schemas.conference.Mutations,
 
 
 class Query(api.schemas.conference.Query,
-            api.schemas.user.Query,
+            api.schemas.auth.Query,
+            api.schemas.profile.Query,
             api.schemas.presentation.Query,
             api.schemas.sponsor.Query,
             graphene.ObjectType):
